services/api/requirements.txt
# ================================================================
# API Service Requirements - Simplified Dependencies
# ================================================================

# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic[email]==2.10.3
python-multipart==0.0.20
jinja2==3.1.4

# Database & ORM
psycopg[binary]==3.2.10
asyncpg==0.29.0
SQLAlchemy==2.0.36
alembic==1.14.0
geoalchemy2==0.15.0
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2_cffi==23.1.0

# HTTP Client
httpx==0.28.1

# Redis & Caching
redis==5.2.1

# Object Storage (MinIO)
minio==7.2.8

# Monitoring & Metrics
prometheus-client==0.21.1

# Rate Limiting
slowapi==0.1.9

# Background Tasks
arq==0.26.1

# Logging
python-json-logger==3.2.1

# Image Processing
Pillow==10.4.0

# File Type Detection
python-magic==0.4.27

# Machine Learning (for fraud detection)
scikit-learn==1.4.2

# Additional Utilities
aiofiles==24.1.0

